Section 13. Capstone - Bug Fixer (Just Code)/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Bug Fixing using crewAI"
authors = [
    { name = "DeepLens Tech", email = "support@deeplens.tech" },
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "aws-opentelemetry-distro>=0.14.2",
    "boto3>=1.40.35",
    "crewai-tools[mcp]>=0.76.0",
    "crewai[mcp,tools]>=0.186.1",
    "litellm>=1.80.0",
    "openinference-instrumentation-crewai>=0.1.13",
    "openinference-instrumentation-litellm>=0.1.25",
    "openlit>=1.36.1",
    "opentelemetry-api",
    "opentelemetry-exporter-otlp",
    "opentelemetry-sdk",
    "markdown>=3.7",
    "rich>=14.3.1",
    "termcolor>=3.3.0",
    "langfuse>=3.14.5",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
